refactor

+++ 2semester/SAOD/12_lab/include/Search.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 template <typename T> long long BinarySearch1(const T &arr, std::size_t len, const T &x, std::size_t &count_of_compares) {
     count_of_compares = 0;
@@ -10,7 +11,7 @@ template <typename T> long long BinarySearch1(const T &arr, std::size_t len, con
         std::size_t middle = (left + right) / 2;
         count_of_compares++;
         if (arr[middle] == x)
-            return (long long)middle;
+            return (int64_t)middle;
 
         if (arr[middle] < x)
             left = middle + 1;
@@ -22,7 +23,7 @@ template <typename T> long long BinarySearch1(const T &arr, std::size_t len, con
     return -1;
 }
 
-template <typename T> long long BinarySearch2(const T &arr, std::size_t len, const T &x, std::size_t &count_of_compares) {
+template <typename T> int64_t BinarySearch2(const T &arr, std::size_t len, const T &x, std::size_t &count_of_compares) {
     count_of_compares = 0;
     std::size_t left = 0, right = len - 1;
 
@@ -37,6 +38,6 @@ template <typename T> long long BinarySearch2(const T &arr, std::size_t len, con
 
     count_of_compares++;
     if (arr[right] == x)
-        return (long long)right;
+        return static_cast<int64_t>(right);
     return -1;
 }
